split surrogate and optimization
